% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmaker_example.R
\name{matchmaker_example}
\alias{matchmaker_example}
\title{show the path to a matchmaker example file}
\usage{
matchmaker_example(name = NULL)
}
\arguments{
\item{name}{the name of a matchmaker example file}
}
\value{
a path to a matchmaker example file
}
\description{
show the path to a matchmaker example file
}
\examples{
matchmaker_example() # list all of the example files

# read in example spelling dictionary
sd <- matchmaker_example("spelling-dictionary.csv")
read.csv(sd, stringsAsFactors = FALSE)

# read in example coded data
coded_data <- matchmaker_example("coded-data.csv")
coded_data <- read.csv(coded_data, stringsAsFactors = FALSE)
str(coded_data)
coded_data$date <- as.Date(coded_data$date)
}
\author{
Zhian N. Kamvar
}
